auto-content-creator/requirements.txt
pyautogen==0.3.0
requests==2.31.0
python-dotenv==1.0.0
web3==7.2.0
solidity-parser==0.1.1

# Additional dependencies that might be useful
pytest==8.3.3
black==24.8.0
flake8==7.1.1